requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.7
pandas==2.1.3
apache-airflow-providers-postgres==5.6.0